Allow to disable Search for users
This plugin can be used to provide a Lucene index for other plugins
that require one. However, in such cases administrators might not
want to use the Search implementation this plugin provides.

To allow for such a case, this commit adds an option to disable
the Search implementation, while still keeping the Lucene index.