Support Java 11 and newer Jenkins versions
Lucene search did not support new Jenkins versions, since the Lucene
version did not support Java 11.

This change update Lucene to 8.11.2, thereby supporting Java 11. It
also bumps the Jenkins version to 2.361.4.

BREAKING CHANGE: Existing users will have to delete the existing index
and rebuild it, since the update from Lucene 3 to 8 does not provide
an upgrade path.

This change also removed the solr-dependency. Since Solr did not support
Jetty 10 (will be added with the next release), it would not be supported
by newer Jenkins version. Solr-support was anyway not included in the
existing releases.